requirements.txt
-e .
aws_cdk.aws_s3==1.102.0
aws_cdk.aws_glue==1.102.0
aws_cdk.aws_iam==1.102.0
aws_cdk.aws_athena==1.102.0
aws_cdk.aws_kinesisfirehose==1.102.0
aws_cdk.aws_kinesis==1.102.0
aws_cdk.aws_rds==1.102.0
aws_cdk.aws_ec2==1.102.0
aws_cdk.aws_dms==1.102.0
aws_cdk.aws_emr==1.102.0
aws_cdk.aws_redshift==1.102.0
aws_cdk.aws_mwaa==1.102.0
aws_cdk.aws_s3_assets==1.102.0
aws_cdk.aws_s3_deployment==1.102.0
aws_cdk.aws_lambda==1.102.0
black==20.8b1
pre-commit==2.9.3
fake-web-events
boto3
black
pre-commit
flake8